[refactor]: rename point_io_ascii module class file names snake_case
