+++ spatialetl/point/io/ascii/default_multi_point_writer.py
